a lot of changes

+++ Desktop/Spring 2020/SWER 313/SWER 313 Project/soap/src/main/java/edu/bu/soap/useraccounts/UserAccountsController.java
@@ -1,13 +1,10 @@
 package edu.bu.soap.useraccounts;
 
-
 import java.io.IOException;
 import java.sql.Date;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.TimeZone;
 
 import javax.servlet.http.HttpServletRequest;
@@ -16,6 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -27,10 +25,6 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 
 //mark class as Controller	
 @RestController
@@ -45,41 +39,40 @@ public class UserAccountsController {
 	 * @Autowired UserAccountsService userAccountsService;
 	 */
 
-//creating a get mapping that retrieves all the userAccounts detail from the database
-	@GetMapping("/useraccounts")
+//creating a get mapping that retrieves the userAccount details from the database
+	@GetMapping("/myprofile")
 	private String getAllUserAccounts() {
-		List<UserAccounts> userAccounts = new ArrayList<UserAccounts>();
-		userAccountsService.getAllUserAccounts().forEach(userAccounts1 -> userAccounts.add(new UserAccounts(userAccounts1.getUserName(),userAccounts1.getUserPassword(),userAccounts1.getBirthDate(),userAccounts1.getEmail(),userAccounts1.getPhotoID(),userAccounts1.getCreationDateTime())));
-		return userAccounts.toString();	
-	}
-
-//creating a get mapping that retrieves the detail of a specific userAccount
-	@GetMapping("/useraccount/{userAccountUserName}")
-	private UserAccounts getUserAccounts(@PathVariable("userAccountUserName") String userAccountUserName) {	
-		return userAccountsService.getUserAccountsByUserName(userAccountUserName);	
+		org.springframework.security.core.Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		String name = auth.getName();
+		return userAccountsService.getUserAccountsByUserName(name).toString();
 	}
 
 //creating a delete mapping that deletes a specified userAccount
 	@DeleteMapping("/useraccount/{userAccountUserName}")
 	private String deleteUserAccount(@PathVariable("userAccountUserName") String userAccountUserName) {
 		try {
-		userAccountsService.delete(userAccountUserName);
-		return "Deleted successfully.";
-		}
-		catch (Exception e) {
+			org.springframework.security.core.Authentication auth = SecurityContextHolder.getContext()
+					.getAuthentication();
+			String name = auth.getName();
+			if (name.equals(userAccountUserName)) {
+				userAccountsService.delete(userAccountUserName);
+				return "Deleted successfully.";
+			} else {
+				return "You can't delete an account that is not yours";
+			}
+		} catch (Exception e) {
 			return "Make sure to enter a right username.";
 		}
 	}
 
-//creating post mapping that post the userAccount detail in the database
+//creating post mapping that register a user and saves his account details in the database
 	@RequestMapping(value = "/registration", method = RequestMethod.POST)
 	public ResponseEntity<String> registerUser(@RequestParam("userName") String name,
 			@RequestParam("userPassword") String password, @RequestParam("email") String email,
 			@RequestParam("birthDate") Date birthDate, @RequestParam("userPhoto") MultipartFile file,
 			HttpServletRequest request, HttpServletResponse response) {
-		String fileDownloadUri = null;
 		byte[] photo = null;
-		String fileName=null;
+		String fileName = null;
 		if (!file.isEmpty()) {
 			try {
 				fileName = StringUtils.cleanPath(file.getOriginalFilename());
@@ -89,11 +82,11 @@ public class UserAccountsController {
 
 				System.out.println("error loading photo");
 			}
-		} else  return new ResponseEntity<String>("Make sure to upload your photo.",HttpStatus.BAD_REQUEST);
+		} else
+			return new ResponseEntity<String>("Make sure to upload your photo.", HttpStatus.BAD_REQUEST);
 		if (email.length() == 0 || name.length() == 0 || password.length() == 0 || birthDate == null) {
-			 return new ResponseEntity<String>("Make sure all field are filled.",HttpStatus.BAD_REQUEST);
-		} else 
-		if (!name.matches("^[a-zA-Z0-9._-]{3,16}$")) {
+			return new ResponseEntity<String>("Make sure all field are filled.", HttpStatus.BAD_REQUEST);
+		} else if (!name.matches("^[a-zA-Z0-9._-]{3,16}$")) {
 			return new ResponseEntity<String>(
 					"Username should be 4 letters minimum, 14 letters maximum. The letters must be alphanumeric.",
 					HttpStatus.BAD_REQUEST);
@@ -106,8 +99,6 @@ public class UserAccountsController {
 		}
 		password = new SecurityConfiguration().getPasswordEncoder().encode(password);
 		UserAccounts userAccounts = new UserAccounts(name, password, birthDate, fileName, photo, email);
-		fileDownloadUri = ServletUriComponentsBuilder.fromCurrentContextPath().path("/downloadFile/").path(name)
-				.toUriString();
 		userAccountsService.saveOrUpdate(userAccounts);
 
 		return new ResponseEntity<String>("You have been registred successfully " + userAccounts.getUserName() + "!!",
@@ -116,82 +107,106 @@ public class UserAccountsController {
 		// return fileDownloadUri;
 	}
 
-//creating put mapping that updates the userAccount detail
-	@PutMapping("/changepassword/{userAccountUserName}")
+//creating put mapping that updates the user account's password
+	@PutMapping("/useraccounts/changepassword/{userAccountUserName}")
 	private ResponseEntity<String> updatePassword(@PathVariable("userAccountUserName") String userAccountUserName,
 			@RequestBody String password) {
 		try {
-		UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
-		if (!password.matches("(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=])(?=\\S+$).{8,30}")) {
-			return new ResponseEntity<String>(
-					"Your password isn't strong enough. It should have at least one special character, one digit, one lowercase and one uppercase character, and it's length is 6 letters minimum, 16 letters maximum.",
-					HttpStatus.BAD_REQUEST);
-		} else {
-			userAccount.setUserPassword(password);
-			userAccountsService.saveOrUpdate(userAccount);
-			return new ResponseEntity<String>("You have changed your password successfully!", HttpStatus.OK);
-		}
-		} catch(Exception e) {
-			return new ResponseEntity<String>("Make sure to enter a right username.", HttpStatus.BAD_REQUEST);
+			org.springframework.security.core.Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+			String name = auth.getName();
+			if (name.equals(userAccountUserName)) {
+			UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
+			if (!password.matches("(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=])(?=\\S+$).{8,30}")) {
+				return new ResponseEntity<String>(
+						"Your password isn't strong enough. It should have at least one special character, one digit, one lowercase and one uppercase character, and it's length is 6 letters minimum, 16 letters maximum.",
+						HttpStatus.BAD_REQUEST);
+			} else {
+				userAccount.setUserPassword(password);
+				userAccountsService.saveOrUpdate(userAccount);
+				return new ResponseEntity<String>("You have changed your password successfully!", HttpStatus.OK);
+			}
+			}
+			else 
+				return new ResponseEntity<String>("Make sure to enter your right username.", HttpStatus.BAD_REQUEST);
+		} catch (Exception e) {
+			return new ResponseEntity<String>("Make sure to enter your right username.", HttpStatus.BAD_REQUEST);
 		}
 	}
-		
 
-	// creating put mapping that updates the userAccount detail
-	@PutMapping("/changeemail/{userAccountUserName}")
+	// creating put mapping that updates the user account's email
+	@PutMapping("/useraccounts/changeemail/{userAccountUserName}")
 	private ResponseEntity<String> updateEmail(@PathVariable("userAccountUserName") String userAccountUserName,
 			@RequestBody String email) {
 		try {
-		UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
-		if (!email.matches("^(.+)@(.+)$")) {
-			return new ResponseEntity<String>("Make sure that you entered a valid email", HttpStatus.BAD_REQUEST);
-		} else {
-		userAccount.setEmail(email);
-		userAccountsService.saveOrUpdate(userAccount);
-		return new ResponseEntity<String>("Email changed successfully to " + email, HttpStatus.OK);
-		}
-		} catch(Exception e) {
-			return new ResponseEntity<String>("Make sure to enter a right username.", HttpStatus.BAD_REQUEST);
+			org.springframework.security.core.Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+			String name = auth.getName();
+			if (name.equals(userAccountUserName)) {
+			UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
+			if (!email.matches("^(.+)@(.+)$")) {
+				return new ResponseEntity<String>("Make sure that you entered a valid email", HttpStatus.BAD_REQUEST);
+			} else {
+				userAccount.setEmail(email);
+				userAccountsService.saveOrUpdate(userAccount);
+				return new ResponseEntity<String>("Email changed successfully to " + email, HttpStatus.OK);
+			}
+			}
+			else 
+				return new ResponseEntity<String>("Make sure to enter your right username.", HttpStatus.BAD_REQUEST);
+		} catch (Exception e) {
+			return new ResponseEntity<String>("Make sure to enter your right username.", HttpStatus.BAD_REQUEST);
 		}
 	}
 
-	// creating put mapping that updates the userAccount detail
-	@PutMapping("/changeuserphoto/{userAccountUserName}")
+	// creating put mapping that updates the user account's photo detail
+	@PutMapping("/useraccounts/changeuserphoto/{userAccountUserName}")
 	private ResponseEntity<String> updateUserPhoto(@PathVariable("userAccountUserName") String userAccountUserName,
 			@RequestParam("userPhoto") MultipartFile file) {
 		try {
-		UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
-		byte[] photo;
-		try {
-			photo = file.getBytes();
-			
-			userAccount.setUserPhoto(photo);
-			userAccountsService.saveOrUpdate(userAccount);
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-		
-		return new ResponseEntity<String>("Photo is uploaded successfully!!", HttpStatus.OK);
-		} catch(Exception e) {
-			return new ResponseEntity<String>("Make sure to enter a right username.", HttpStatus.BAD_REQUEST);
+			org.springframework.security.core.Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+			String name = auth.getName();
+			if (name.equals(userAccountUserName)) {
+			UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
+			byte[] photo;
+			try {
+				photo = file.getBytes();
+
+				userAccount.setUserPhoto(photo);
+				userAccountsService.saveOrUpdate(userAccount);
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+
+			return new ResponseEntity<String>("Photo is uploaded successfully!!", HttpStatus.OK);
+			}
+			else 
+				return new ResponseEntity<String>("Make sure to enter your right username.", HttpStatus.BAD_REQUEST);
+		} catch (Exception e) {
+			return new ResponseEntity<String>("Make sure to enter your right username.", HttpStatus.BAD_REQUEST);
 		}
 	}
 
-	// creating put mapping that updates the userAccount detail
-	@PutMapping("/changebirthdate/{userAccountUserName}")
+	// creating put mapping that updates the user account's birth date
+	@PutMapping("/useraccounts/changebirthdate/{userAccountUserName}")
 	private ResponseEntity<String> updateBirthDate(@PathVariable("userAccountUserName") String userAccountUserName,
 			@RequestBody String birthDate) throws ParseException {
 		try {
-			TimeZone.setDefault(TimeZone.getTimeZone("UTC")); //to get the accurate date
+			org.springframework.security.core.Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+			String name = auth.getName();
+			if (name.equals(userAccountUserName)) {
+			TimeZone.setDefault(TimeZone.getTimeZone("UTC")); // to get the accurate date
 			DateFormat df = new SimpleDateFormat("yyyy-MM-dd");
 			java.util.Date date = (java.util.Date) df.parse(birthDate);
 
-		UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
-		userAccount.setBirthDate(date);
-		userAccountsService.saveOrUpdate(userAccount);
-		return new ResponseEntity<String>("You have changed your birth date successfully!", HttpStatus.OK);
-		} catch(Exception e) {
-			return new ResponseEntity<String>("Make sure to enter a right username and a right date format.", HttpStatus.BAD_REQUEST);
+			UserAccounts userAccount = userAccountsService.getUserAccountsByUserName(userAccountUserName);
+			userAccount.setBirthDate(date);
+			userAccountsService.saveOrUpdate(userAccount);
+			return new ResponseEntity<String>("You have changed your birth date successfully!", HttpStatus.OK);
+			}
+			else 
+				return new ResponseEntity<String>("Make sure to enter your right username.", HttpStatus.BAD_REQUEST);
+		} catch (Exception e) {
+			return new ResponseEntity<String>("Make sure to enter a right username and a right date format.",
+					HttpStatus.BAD_REQUEST);
 		}
 	}
